requirements.txt
streamlit
langchain
transformers
diffusers
torch
Pillow
huggingface_hub
python-dotenv


